refactor: postfixNotation.js 확장자 ts로 변경
